Add the Select New Grid, EventListener, and Hover
Adding the new function that clears the grid, crea
te a loop for the number of rows and cell. Click
event on the button that will generate a prompt
that is valid between 1 and 100 and will generate
the grid. in the event calling the grid function and
the Hover function so the cells will change change
color once you hover over them.